extracted menu to other component
